requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==2.0.2
anyio==3.7.0
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.97.0
fastapi-mail==1.3.1
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.89.0
google-auth==2.20.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
h11==0.14.0
httpcore==0.17.2
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.3.1
multidict==6.0.4
nodeenv==1.8.0
oauthlib==3.2.2
openai==0.28.1
passlib==1.7.4
prisma==0.10.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pymongo==4.5.0
pyparsing==3.0.9
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.11.8
tqdm==4.66.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
